% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_team.R
\name{football_team}
\alias{football_team}
\title{Football team in England}
\usage{
football_team(london, manchestar, newcastle)
}
\arguments{
\item{london}{most of the team in}

\item{manchestar}{major two significant team in}

\item{newcastle}{rising team}
}
\description{
Football team in England
}
