% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_quotes.R
\name{rat_quotes}
\alias{rat_quotes}
\title{Title Rat Quotes}
\usage{
rat_quotes(rat)
}
\arguments{
\item{rat}{just type in ur name or something}
}
\value{
rat quotes of mine
}
\description{
Title Rat Quotes
}
