% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy.R
\name{msy}
\alias{msy}
\title{Calculate maxium sustainable yield}
\usage{
msy(k, r)
}
\arguments{
\item{k}{is the carrying capacity(individuals)}

\item{r}{is intrinsic grouwth rate (1/yr)}
}
\value{
returns the maximun sustainable harvest for a fishery
}
\description{
Calculate maxium sustainable yield
}
\examples{
msy(300,0.06)
}
